Fixed a bug in the post post invocation and the EventsRepository

+++ backend/src/main/java/com/goodtimes/events/EventsRepository.java
@@ -1,11 +1,12 @@
 package com.goodtimes.events;
 
-import com.goodtimes.events.GoodtimeEvent;
 import org.springframework.data.mongodb.repository.MongoRepository;
+
+import java.math.BigInteger;
 
 /**
  * Created by rockem on 16/2/15.
  */
-public interface EventsRepository extends MongoRepository<GoodtimeEvent, Long> {
+public interface EventsRepository extends MongoRepository<GoodtimeEvent, BigInteger> {
 
 }
